+++ server/src/main/java/com/seniclass/server/domain/lecture/service/LectureServiceImpl.java
@@ -1,0 +1,90 @@
+package com.seniclass.server.domain.lecture.service;
+
+import com.seniclass.server.domain.category.domain.SubCategory;
+import com.seniclass.server.domain.category.repository.SubCategoryRepository;
+import com.seniclass.server.domain.lecture.domain.Lecture;
+import com.seniclass.server.domain.lecture.dto.request.LectureCreateRequest;
+import com.seniclass.server.domain.lecture.dto.request.LectureUpdateRequest;
+import com.seniclass.server.domain.lecture.dto.response.LectureResponse;
+import com.seniclass.server.domain.lecture.repository.LectureRepository;
+import com.seniclass.server.global.exception.CommonException;
+import com.seniclass.server.global.exception.errorcode.LectureErrorCode;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class LectureServiceImpl implements LectureService {
+
+    private final LectureRepository lectureRepository;
+    private final SubCategoryRepository subCategoryRepository;
+
+    @Override
+    public LectureResponse createLecture(LectureCreateRequest request) {
+
+        SubCategory subCategory =
+                subCategoryRepository
+                        .findById(request.subCategoryId())
+                        .orElseThrow(
+                                () -> new CommonException(LectureErrorCode.SUB_CATEGORY_NOT_FOUND));
+        Integer cohort =
+                lectureRepository
+                        .findTopByNameOrderByCohortDesc(request.name())
+                        .map(lecture -> lecture.getCohort() + 1)
+                        .orElse(1);
+
+        Lecture lecture =
+                Lecture.createLecture(
+                        subCategory,
+                        request.name(),
+                        cohort,
+                        request.level(),
+                        request.startDate(),
+                        request.endDate(),
+                        request.maxStudent(),
+                        request.fee(),
+                        request.instruction(),
+                        request.description());
+
+        return LectureResponse.from(lectureRepository.save(lecture));
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public LectureResponse getLecture(Long lectureId) {
+        Lecture lecture = getLectureEntity(lectureId);
+        return LectureResponse.from(lecture);
+    }
+
+    @Override
+    public LectureResponse updateLecture(Long lectureId, LectureUpdateRequest request) {
+        Lecture lecture = getLectureEntity(lectureId);
+        SubCategory subCategory =
+                subCategoryRepository
+                        .findById(request.subCategoryId())
+                        .orElseThrow(
+                                () -> new CommonException(LectureErrorCode.SUB_CATEGORY_NOT_FOUND));
+        lecture.updateLecture(
+                subCategory,
+                request.level(),
+                request.startDate(),
+                request.endDate(),
+                request.maxStudent(),
+                request.instruction(),
+                request.description());
+
+        return LectureResponse.from(lecture);
+    }
+
+    public void deleteLecture(Long lectureId) {
+        lectureRepository.deleteById(lectureId);
+    }
+
+    private Lecture getLectureEntity(Long lectureId) {
+        return lectureRepository
+                .findById(lectureId)
+                .orElseThrow(() -> new CommonException(LectureErrorCode.LECTURE_NOT_FOUND));
+    }
+}
